Changes: - Modified the amount field to double.
ToDo
- Read Column Names without explicitly hardcoding the column names to the FileItemReader.
- Link all transactions to the associated accounts.
- Read from Directory without hardcoding the absolute path.

+++ src/main/java/io/vultr/cld/paycons/domain/TxDto.java
@@ -7,8 +7,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.math.BigDecimal;
-
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
@@ -22,5 +20,5 @@ public class TxDto extends AbstractJsonUtil {
     private String time;        //Transaction Time
     private String descr;       //Transaction Description
     private Character type;     //Transaction Type -> Debit or Credit.
-    private BigDecimal amount;  //Transaction AMount
+    private Double amount;  //Transaction AMount
 }
